+++ R/url_ejamapi.R
@@ -1,22 +1,23 @@
 
 
 #' Get URL(s) of HTML summary reports for use with EJAM-API
+#' @seealso [ejamapi()]
 #' @details
-#' - Work in progress - initial draft relied on API from
+#' - This is work in progress to some extent -- this and the API may be add features in later releases.
+#'
+#' - Relies on API from
 #'   https://github.com/edgi-govdata-archiving/EJAM-API
 #'
-#'   (see parameter `baseurl` that used the /report endpoint)
-#'
-#' - Another option in the future might be to construct a URL that is a link to the live EJAM
+#' - Another option in the future might be to construct a URL that is a "deep link" to the live EJAM
 #'   app but has url-encoded parameters that are app settings, such as sitepoints, radius_default, etc.
 #'
 #' - Will try to use the same input parameters as [ejamit()] does.
 #'
-#' - The 12/2025 API used ejam2report() with these parameter settings:
+#' - The API as of mid-2026 used `ejam2report()` with these parameter settings:
 #'   - `sitenumber = 1`
 #'   - `report_title="EJSCREEN Community Report"`
 #'
-#'   So it was not yet accepting parameters used by [ejamit()] and [ejam2report()] such as
+#'   Therefore, it was not yet accepting parameters used by [ejamit()] and [ejam2report()] such as
 #'   - sitenumber=0 (for a multisite report)
 #'   - logo_path
 #'   - report_title
@@ -28,7 +29,7 @@
 #' @param lat,lon can be provided as vectors of coordinates instead of providing sitepoints table
 #' @param radius  see [ejamit()], default is 0 if fips or shapefile specified
 #'
-#' @param fips  see [ejamit()] but this initial version only works for a blockgroup FIPS!
+#' @param fips  see [ejamit()]
 #'
 #' @param shapefile  see [ejamit()], but each polygon is encoded as geojson string
 #'   which might get too long for encoding in a URL for the API using GET
@@ -37,7 +38,7 @@
 #' @param as_html Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example
 #' @param linktext used as text for hyperlinks, if supplied and as_html=TRUE
 #' @param ifna URL shown for missing, NA, NULL, bad input values
-#' @param baseurl do not change unless endpoint actually changed
+#' @param baseurl do not change unless endpoint actually changed. See [ejamapi()] for a better way to handle choice of endpoint.
 #'
 #' @param sitenumber
 #'